@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.9",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "~11.2.9",
-    "@angular/compiler": "~11.2.9",
-    "@angular/core": "~11.2.9",
-    "@angular/forms": "~11.2.9",
+    "@angular/animations": "~14.0.6",
+    "@angular/cdk": "^14.0.5",
+    "@angular/common": "~14.0.6",
+    "@angular/compiler": "~14.0.6",
+    "@angular/core": "~14.0.6",
+    "@angular/forms": "~14.0.6",
     "@angular/material": "^11.2.13",
-    "@angular/platform-browser": "~11.2.9",
-    "@angular/platform-browser-dynamic": "~11.2.9",
-    "@angular/router": "~11.2.9",
+    "@angular/platform-browser": "~14.0.6",
+    "@angular/platform-browser-dynamic": "~14.0.6",
+    "@angular/router": "~14.0.6",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.3"
@@ -28,7 +28,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1102.8",
     "@angular/cli": "~11.2.8",
-    "@angular/compiler-cli": "~11.2.9",
+    "@angular/compiler-cli": "~14.0.6",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
@@ -42,6 +42,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.7.4"
   }
 }
